+++ CHANGELOG.md
@@ -9,3 +9,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Initial release of wavelink-cli
+
+## [0.0.2] - 2026-01-11
+
+### Fixed
+- Resolved build issues.
+
+## [0.0.3] - 2026-01-11
+
+### Changed
+- Updated `mix set-output` to remove unassigned devices from the mix instead of reassigning them to another mix.
+- Removed the requirement for at least two mixes when using `mix set-output`.
+